Update run_pipeline.py
attempting to make the pipeline runner more robust. in the previous version of this script, entire runs could crash out if a single step failed for a single file. this will (hopefully!) bypass this so that the pipeline keeps running even if a single item tries to gum up the works