v0.1.95
A couple of things. Hammered out the last pieces of the vocal acoustics feature extraction (buyer beware: again, I'm not an expert in this). Hardened the script a bit to add in some silence in places where necessary so that short chunks of audio won't completely fail. Integrated all of this into the `conversation_video` pipeline.